+++ src/api/content/qna.document.library/qna.document.library.controller.ts
@@ -1,73 +1,33 @@
-/* eslint-disable linebreak-style */
-/* eslint-disable no-multiple-empty-lines */
-/* eslint-disable linebreak-style */
-/* eslint-disable padded-blocks */
-/* eslint-disable indent */
-/* eslint-disable no-console */
-/* eslint-disable @typescript-eslint/no-unused-vars */
-/* eslint-disable linebreak-style */
 import express from 'express';
 import { ResponseHandler } from '../../../common/handlers/response.handler';
 import { QnaDocumentLibraryValidator } from './qna.document.library.validator';
-import { BaseController } from '../../base.controller';
 import { ErrorHandler } from '../../../common/handlers/error.handler';
-
 import { uuid } from '../../../domain.types/miscellaneous/system.types';
-import { QnaDocumentLibrary } from '../../../database/models/qna.document/qna.document.library.model';
 import { QnaDocumentLibraryService } from '../../../database/services/content/qna.document.library.service';
 import { QnaDocumentLibraryCreateModel } from '../../../domain.types/content/qna.document.library.domain.types';
 import { QnaDocumentLibraryUpdateModel } from '../../../domain.types/content/qna.document.library.domain.types';
-// import { validateParamAsUUID } from './base.validator.ts';
 
 ///////////////////////////////////////////////////////////////////////////////////////
 
 export class QnaDocumentLibraryController {
-    //#region member variables and constructors
-
+    
     public _service: QnaDocumentLibraryService;
 
     _validator: QnaDocumentLibraryValidator = new QnaDocumentLibraryValidator();
 
     constructor() {
-        // super();
         this._service = new QnaDocumentLibraryService();
     }
 
-    //#endregion
-
-    getAll = async (request: express.Request, response: express.Response) => {
-        try {
-            const record = await this._service.getAll();
-            const message = 'Qna-Document-Library retrieved successfully!';
-            return ResponseHandler.success(request, response, message, 200, record);
-        } catch (error) {
-            ResponseHandler.handleError(request, response, error);
-        }
-    };
-
     create = async (request: express.Request, response: express.Response) => {
-        console.log('Create request call');
         try {
-            // await this.authorize('QnaDocumentVersion.Create', request, response);
             var model: QnaDocumentLibraryCreateModel = await this._validator.validateCreateRequest(request);
             const record = await this._service.create(model);
             if (record === null) {
                 ErrorHandler.throwInternalServerError('Unable to add qna document library!');
             }
-            const message = 'QnaDocumentLibrary added successfully!';
+            const message = 'Qna Document Library added successfully!';
             return ResponseHandler.success(request, response, message, 201, record);
-        } catch (error) {
-            ResponseHandler.handleError(request, response, error);
-        }
-    };
-
-    getById = async (request: express.Request, response: express.Response) => {
-        try {
-            // await this.authorize('QnaDocumentVersion.GetById', request, response);
-            var id: uuid = await this._validator.validateParamAsUUID(request, 'id');
-            const record = await this._service.getById(id);
-            const message = 'QnaDocumentLibrary retrieved successfully!';
-            return ResponseHandler.success(request, response, message, 200, record);
         } catch (error) {
             ResponseHandler.handleError(request, response, error);
         }
@@ -75,28 +35,57 @@ export class QnaDocumentLibraryController {
 
     update = async (request: express.Request, response: express.Response) => {
         try {
-            // await this.authorize('QnaDocumentVersion.Update', request, response);
             const id = await this._validator.validateParamAsUUID(request, 'id');
             var model: QnaDocumentLibraryUpdateModel = await this._validator.validateUpdateRequest(request);
             const updatedRecord = await this._service.update(id, model);
-            const message = 'QnaDocumentLibrary updated successfully!';
+            const message = 'Qna Document Library updated successfully!';
             ResponseHandler.success(request, response, message, 200, updatedRecord);
         } catch (error) {
             ResponseHandler.handleError(request, response, error);
         }
     };
 
-    delete = async (request: express.Request, response: express.Response): Promise<void> => {
+    getById = async (request: express.Request, response: express.Response) => {
         try {
-            // await this.authorize('QnaDocument.Delete', request, response);
             var id: uuid = await this._validator.validateParamAsUUID(request, 'id');
-            const result = await this._service.delete(id);
-            const message = 'QnaDocumentLibrary deleted successfully!';
-            ResponseHandler.success(request, response, message, 200, result);
+            const record = await this._service.getById(id);
+            if (record != null) {
+                const message = 'Qna Document Library retrieved successfully!';
+                return ResponseHandler.success(request, response, message, 200, record);
+            } else {
+                const message = 'Qna Document Library Not found';
+                return ResponseHandler.failure(request, response, message, 404);
+            }
         } catch (error) {
             ResponseHandler.handleError(request, response, error);
         }
     };
 
-    
+    getAll = async (request: express.Request, response: express.Response) => {
+        try {
+            const record = await this._service.getAll();
+            const message = 'Qna Document Library retrieved successfully!';
+            return ResponseHandler.success(request, response, message, 200, record);
+        } catch (error) {
+            ResponseHandler.handleError(request, response, error);
+        }
+    };
+
+    delete = async (request: express.Request, response: express.Response): Promise<void> => {
+        var id: uuid = await this._validator.validateParamAsUUID(request, 'id');
+        const record = await this._service.getById(id);
+        if (record != null) {
+            try {
+                const result = await this._service.delete(id);
+                const message = 'Qna Document Library deleted successfully!';
+                ResponseHandler.success(request, response, message, 200, result);
+            } catch (error) {
+                ResponseHandler.handleError(request, response, error);
+            }
+        } else {
+            const message = 'Qna Document Library Not found';
+            ResponseHandler.failure(request, response, message, 404);
+        }
+    };
+
 }
