+++ src/api/content/qna.document/qna.document.controller.ts
@@ -1,59 +1,29 @@
-/* eslint-disable padded-blocks */
-/* eslint-disable indent */
-/* eslint-disable no-console */
-/* eslint-disable @typescript-eslint/no-unused-vars */
-/* eslint-disable linebreak-style */
 import express from 'express';
 import { ResponseHandler } from '../../../common/handlers/response.handler';
 import { QnaDocumentValidator } from './qna.document.validator';
-import { BaseController } from '../../base.controller';
-import { QnaDocumentGroupsService } from '../../../database/services/content/qna.document.group.service';
 import { ErrorHandler } from '../../../common/handlers/error.handler';
-import {
-    QnaDocumentGroupCreateModel,
-    QnaDocumentGroupUpdateModel,
-} from '../../../domain.types/content/qna.document.group.domain.types';
 import { uuid } from '../../../domain.types/miscellaneous/system.types';
-import { QnaDocumentGroup } from '../../../database/models/qna.document/qna.document.groups.model';
 import { QnaDocumentService } from '../../../database/services/content/qna.document.service';
 import {
     QnaDocumentCreateModel,
     QnaDocumentSearchFilters,
+    QnaDocumentUpdateModel,
 } from '../../../domain.types/content/qna.document.domain.types';
-import { QnaDocumentUpdateModel } from '../../../domain.types/content/qna.document.domain.types';
-// import { validateParamAsUUID } from './base.validator.ts';
-import { QnaDocument } from '../../../database/models/qna.document/qna.document.model';
 
 ///////////////////////////////////////////////////////////////////////////////////////
 
 export class QnaDocumentController {
-    //#region member variables and constructors
-
+    
     public _service: QnaDocumentService;
 
     _validator: QnaDocumentValidator = new QnaDocumentValidator();
 
     constructor() {
-        // super();
         this._service = new QnaDocumentService();
     }
 
-    //#endregion
-
-    getAll = async (request: express.Request, response: express.Response) => {
-        try {
-            const record = await this._service.getAll();
-            const message = 'Qna-Document retrieved successfully!';
-            return ResponseHandler.success(request, response, message, 200, record);
-        } catch (error) {
-            ResponseHandler.handleError(request, response, error);
-        }
-    };
-
     create = async (request: express.Request, response: express.Response) => {
-        console.log('Create request call');
         try {
-            // await this.authorize('QnaDocumentGroup.Create', request, response);
             var model: QnaDocumentCreateModel = await this._validator.validateCreateRequest(request);
             const record = await this._service.create(model);
             if (record === null) {
@@ -66,21 +36,8 @@ export class QnaDocumentController {
         }
     };
 
-    getById = async (request: express.Request, response: express.Response) => {
-        try {
-            // await this.authorize('QnaDocument.GetById', request, response);
-            var id: uuid = await this._validator.validateParamAsUUID(request, 'id');
-            const record = await this._service.getById(id);
-            const message = 'Qna-Document retrieved successfully!';
-            return ResponseHandler.success(request, response, message, 200, record);
-        } catch (error) {
-            ResponseHandler.handleError(request, response, error);
-        }
-    };
-
     update = async (request: express.Request, response: express.Response) => {
         try {
-            // await this.authorize('QnaDocument.Update', request, response);
             const id = await this._validator.validateParamAsUUID(request, 'id');
             var model: QnaDocumentUpdateModel = await this._validator.validateUpdateRequest(request);
             const updatedRecord = await this._service.update(id, model);
@@ -91,15 +48,57 @@ export class QnaDocumentController {
         }
     };
 
-    delete = async (request: express.Request, response: express.Response): Promise<void> => {
+    getById = async (request: express.Request, response: express.Response) => {
         try {
-            // await this.authorize('QnaDocument.Delete', request, response);
             var id: uuid = await this._validator.validateParamAsUUID(request, 'id');
-            const result = await this._service.delete(id);
-            const message = 'Qna-Document deleted successfully!';
-            ResponseHandler.success(request, response, message, 200, result);
+            const record = await this._service.getById(id);
+            if (record != null) {
+                const message = 'Qna-Document retrieved successfully!';
+                return ResponseHandler.success(request, response, message, 200, record);
+            } else {
+                const message = 'Qna-Document not found';
+                return ResponseHandler.failure(request, response, message, 404);
+            }
         } catch (error) {
             ResponseHandler.handleError(request, response, error);
+        }
+    };
+
+    getAll = async (request: express.Request, response: express.Response) => {
+        try {
+            const record = await this._service.getAll();
+            const message = 'Qna-Document retrieved successfully!';
+            return ResponseHandler.success(request, response, message, 200, record);
+        } catch (error) {
+            ResponseHandler.handleError(request, response, error);
+        }
+    };
+
+    getByStatus = async (request: express.Request, response: express.Response) => {
+        try {
+            const status: boolean = request.params.status === 'true';
+            const records = await this._service.getByStatus(status);
+            const message = 'Qna document By status retrieved successfully!';
+            return ResponseHandler.success(request, response, message, 200, records);
+        } catch (error) {
+            ResponseHandler.handleError(request, response, error);
+        }
+    };
+
+    delete = async (request: express.Request, response: express.Response): Promise<void> => {
+        var id: uuid = await this._validator.validateParamAsUUID(request, 'id');
+        const record = await this._service.getById(id);
+        if (record != null) {
+            try {
+                const result = await this._service.delete(id);
+                const message = 'Qna-Document deleted successfully!';
+                ResponseHandler.success(request, response, message, 200, result);
+            } catch (error) {
+                ResponseHandler.handleError(request, response, error);
+            }
+        } else {
+            const message = 'Qna-Document not found';
+            ResponseHandler.failure(request, response, message, 404);
         }
     };
 
@@ -113,4 +112,5 @@ export class QnaDocumentController {
             ResponseHandler.handleError(request, response, error);
         }
     };
+    
 }
